+++ tests/integration/cli/commands/test_run.py
@@ -216,7 +216,7 @@ def example(tmpdir):
                 """\
                 [build-system]
                 requires = ["setuptools"]
-                backend = "setuptools.build_meta"
+                build-backend = "setuptools.build_meta"
                 """
             )
         )